requirements.txt
attrs==20.1.0
base58==2.0.1
bitarray==1.2.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
cytoolz==0.10.1
decorator==4.4.2
eth-abi==2.1.1
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.4
hexbytes==0.2.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
ipfshttpclient==0.6.1
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.13.0
pycparser==2.20
pycryptodome==3.9.8
pyrsistent==0.16.0
python-decouple==3.3
-e git+git@github.com:Telegram-Totality/python-telegram-bot.git@648ebaea788725ca9eddef951bc9852affd84cc9#egg=python_telegram_bot
requests==2.24.0
rlp==1.2.0
six==1.15.0
toolz==0.10.0
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
varint==1.0.2
web3==5.12.0
websockets==8.1
zipp==3.1.0
